refactor(AMD/CommonJS exported module): Export module name instead of module object
When the module is imported using AMD/CommonJS, it is expected to expose its name instead of the
Angular module object. This is because angular.module() expects the dependencies as string instead
of actual module objects.

Fix #326. Breaking change: code that relies on the module object to be exported (accessing the name
via `.name` for example) will break, since the name is now directly returned.