add the option to use ticksArray as array of objects, with legend and value properties (#662)
same as we do with steps array, but to customize the ticks without limiting the selection steps. 

I know this can be already done with a variety of methods, but AFAIK only with functions. In my case I store these labels in a mongo db, so storing logic is not ideal. Putting those labels in a table of objects without limiting the selection as stepsArray solves both issues, storing and displaying.